FTRepCBR.Workflow.FW/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ftrepcbr-workflow-fw"
dynamic = ["version"]
description = "DQCR Framework for building and validating DQCR projects."
readme = "docs/Framework_Packaging_Changes.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6,<7",
  "Jinja2>=3.1,<4",
]

[project.scripts]
fw2 = "FW.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
FW = "src"

[tool.setuptools.dynamic]
version = {attr = "FW.__version__"}
